requirements.txt
mediapipe>=0.10.9,<0.11
opencv-python>=4.9.0.80
numpy>=1.24.0
PyYAML>=6.0.1
protobuf<5
python-osc>=1.8.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

